RAG/Lab 2/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
tavily-python>=0.5.0
